cli: Move entrypoint code to new package
